+++ common/data/bo/jobTagBO.js
@@ -1,4 +1,7 @@
-export class JobTagBO{
+import { TAG_SOURCE_TYPE_CUSTOM } from "../../index";
+export class JobTagBO {
     jobId;
     tags;
+    sourceType = TAG_SOURCE_TYPE_CUSTOM;
+    source = null;
 }