+++ common/data/domain/tag.js
@@ -1,13 +1,16 @@
 export class Tag {
-    /**
-     * 编号
-     */
-    tagId;
-    /**
-     * 标签名
-     */
-    tagName;
-    createDatetime;
-    updateDatetime;
-  }
-  
+  /**
+   * 编号
+   */
+  tagId;
+  /**
+   * 标签名
+   */
+  tagName;
+  createDatetime;
+  updateDatetime;
+  /**
+  * 0代表不公开;1:代表公开;
+  */
+  isPublic=1;
+}