+++ entrypoints/admin/src/data/JobTagData.tsx
@@ -1,11 +1,14 @@
+import { JobTagDTO } from "@/common/data/dto/jobTagDTO";
+
 export type JobTagData = {
-    id:string;
+    id: string;
     jobId: string;
     name: string;
-    platform:string;
-    companyName:string;
+    platform: string;
+    companyName: string;
     createDatetime: string;
     updateDatetime: string;
     nameArray: string[];
     idArray: string[];
+    tagArray: JobTagDTO[];
 }