+++ CHANGELOG.md
@@ -4,11 +4,13 @@
 
 ### ⭐ Added
 
-1. (ContentScript)记录职位技能标签，职位福利标签
+1. (ContentScript)记录职位技能标签，职位福利标签。
+2. 职位标签新增sourceType(来源类型)，source（来源）。
+3. 标签新增是否公开选项字段。
 
 ### 🐛 Fixed
 
-1. (Sidepanel)修复自动化访问搜索页面时出现的网络错误提示（执行自动化任务前清理网站的cookies（51JOB））
+1. (Sidepanel)修复自动化访问搜索页面时出现的网络错误提示。（执行自动化任务前清理网站的cookies（51JOB））
 
 ### 🔨 Dependency